# app/rest/handler/user/response/login.go
// Resik
// Author: Dedy Fajar Setyawan
// See: https://github.com/dedyf5/resik

package response

type UserCredential struct {
	Username string `json:"username" example:"sakisaki"`
	Token    string `json:"token" example:"xxx.xxx.xxx"`
}